+++ lib/quiz_shower_app.dart
@@ -2,8 +2,8 @@ import 'package:firebase_ui_localizations/firebase_ui_localizations.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:quiz_shower/ui/screen/home_screen/home_screen.dart';
 import 'package:quiz_shower/ui/screen/setting_screen/setting_screen.dart';
-import 'package:quiz_shower/ui/widget/quiz_shower_scaffold.dart';
 
 class QuizShowerApp extends StatelessWidget {
   const QuizShowerApp({super.key});
